:sparkles: graph init vsn
Initialize the VSN subgraph.
Note that we're using the implementation contract for the init phase
so all the events can be retrieved, but we will later change the address
to be the proxy contract (0x699Ccf919C1dfdFa4C374292f42CAdC9899BF753)
once the code is scaffolded.
```
npx graph init vision-vsn vsn \
--from-contract=0xda528Dd3e7F2cd16E5F4175902c64F5993bC297c \
--network=mainnet \
--protocol=ethereum \
--skip-git \
--contract-name=VSN \
--index-events \
--start-block=22688805
```